Bug 692694: unaligned memory access in image ht code.
When image data does not land exactly on the necessary alignment for SSE2
optimisations, we handle the "left over" samples on their own, before the
suitably aligned samples. When we come to copy the halftoned data to the
target device, we first handle the "left over" samples, and then the
SSE2 aligned samples in separate calls to copy_mono/copy_plane. The "left
over" samples never number more than 16.

Previously, after dealing with the non-SSE2 aligned samples, we'd increment
the samples pointer two bytes (to skip the up to 16 bits of unaligned samples).
The problem is, that causes us to 32 bits at time, on 16 bit aligned boundaries,
which on SPARC causes a bus error.

We now use the copy_mono/copy_plane API parameter which gives the function an
offset into the samples which is should use before starting the operation.
Implementations of copy_mono/copy_plane must handle non-aligned samples
gracefully anyway.

No cluster differences.