Allow use of unmanaged color transformations for source DeviceXXX colors
This introduces a CMM that can hand back "links" to gs where the
links in this case contain pointers to the simple procedures that
transform between the various color spaces.  This CMM (or non-CMM) can
be invoked when we encounter DeviceRGB, DeviceGray or DeviceCMYK color
spaces by use of the parameter -dUseFastColor.