Bug 692686: use same scan conversion settings for "clip" as for "fill"
The PDF job in question renders most of its text as simple text render mode 0,
but about three glyphs are done using tr mode 7, and showing an image through
the resulting clip.

The problem is that when we image a glyph, we use fill_adjust 0 (rather than the
default of ~0.3), otherwise glyphs are overly bold. When using one of the clip
text modes in PDF, however, we still used the default fill_adjust, which
resulted in an overly "bold" clip path.

This causes a number of cluster differences: several are progressions, the
others are pixel differences - mostly cases where resolution and/or zoom level
affect whether we precisely match Acrobat.