Do not use hexadecimal names for type42 Charstrings and Encoding
Bug #692711 - When creating CharStrings and Encoding for an embedded
TrueType font, to create a type 42 font, we need to supply numeric names
according to the spec. We were doing this using hex values which has the
unfortunate effect of creating 5 of the names as /A to /F.

These are not numeric! In fact they duplicate 5 of the Ghostscript
standard names. While this is not actually a problem when rendering, it
*is* a problem for ps2write because it uses the names to rebuild the font
for embedding. If the name is one of the standard names then it leads
(through complex routes) to us creating a CMAP subtable where the GID is
not correct, we use the value of the stadnard name instead of the correct
numeric value.

By using decimal numbers we avoid this problem.

This change exhibits progressions in the regression files Bug688421.pdf
(which is now 100% correct with the earlier fixes for composite glyphs)
and Bug691121.pdf