Bug 690779: Provide fallback for missing CIDFonts in PDF
Rather than throw an error when a PDF references a CIDFont that is not
available, we'll now fall back to using DroidSansFallback.ttf, by default.

This gives a reasonable analogue of other PDF consumers' behavior.

Results in extensive differences in Bug692320.pdf because we now complete
content streams, and image the text they contain, rather than skipping
them. NOTE: this does not make the output from Bug692320.pdf "correct", but
arguably makes it "less wrong".

Other than that, no differences.