Fix crash when using a TrueType substitute for a missing CIDFont with pdfwrite
The changes to better handle PCL fonts accidentally left out an important
test which could lead to a seg fault if a PDF file was processed which
contained a reference to a non-embedded CIDFont, and we had defined a
TrueType substitute to use instead.

This restores the old behaviour of exiting with an error.

Bug692320 should no longer seg fault with cluster testing.