Re-enable CDevProc for substituted CIDFonts.
Previously, CDevProc execution was disabled for CIDFonts which had been
substituted by a TTF (through cidfmap, for example) because the index that
gets passed to zchar_set_cache() is the TTF GID, not the CID, and the
CDevProc requires the CID.

We can, however, retrieve the CID from the text enumerator, and pass that to
the CDevProc, which is what this commit implements.

Cluster differences in Bug692320.pdf because we now reposition glyphs with the
CDevProc where we previously were not.

No other differences.