Bug 690779 (pt2): revisions to CIDFont substition.
The original code (commit 18a51701) didn't work well with pdfwrite, so this
revision handles substition by modifying the CIDFont resource .map dictionary.
This means that the "fake" CIDFont is created with the requested parameters
from the start.

Also, add command line paramters and environment variables to control the
path to and file name of the subsitute TTF.

Differences in Bug692320.pdf.

No other cluster differences.