ps2write - Use Font level Resources for type 3 font CharProcs
Part of Bug #691928, file tpc2.ps. It is possible (in PDF) for a font to have
Resources separately from teh page stream. This is properly catered for in
PDF interpreters.

However, ps2write's prolog only allowed resources defined at the page level
when executing CharProc procedures from a type 3 font (because we are in the
page stream at the time, not the font object).

If the type 3 font executes findfont, and the font is not defined at the page
level (often the case for the Quark q font and other substitutes) then the
font cannot be found. Due to a quirk of the way fonts are defined, the *first*
page to use the type 3 font always includes definitions for any fonts used by
the type 3 font, so this problem only exhibits on mutiple page documents.

This was tricky to fix, we cannot simply replace the Resources in the current
Context, as we need to return to the pre-existing Context after running the
CharProc. Additionally, type 3 fonts can execute other type 3 fonts which may
themselves call other fonts. It was neccesary to implement a stack of saved
resources in order to deal with this situation.

Expected Differences
tpc2.ps should now work properly with ps2write