Bug 692754: have configure setup endian setting for lcms
The lcms header files have some educated guesses for the current platform
big/little endianness, but they are not comprehensive (for example, PA-RISC
HP UX defeats it).

So, add a test to our configure scripts, and pass the setting as required as a
compiler option for lcms. NOTE: I added an explicit test for endianness because
the autoconf built-in test has problems in recent releases.

Also, noticed in passing: revise the other explicit compile/link tests to use
the AC_LANG_PROGRAM() macro - currently not doing so produces a warning, but
the implication is future versions will throw an error.

No cluster differences.