Movement of proof profile an device link profile to device.
The proofing profile and the device link profile are now in the device's
profile structure.  This makes it much easier to include in the proofing
profile and/or the device link profile during the link transformation
creation.   This commit includes all the changes except for the actual
creation of the link with the proof.  That will come in a separate commit
since there are a lot of changes in this commit and I want to spend
a little more time on that to make it clean and
dependent upon the capabilities of the CMM.  The reason is that there are
some CMMs that support the chaining together of multiple transformations and
some that do not.  Those that do not may not be able to handle both a
proofing profile AND a device link profile.  We want to handle that case
gracefully.