Proof and DeviceLink ICC profile support
This enables the use of the device link and/or a proofing profile.   When
present, the transformations are as follows
src profile -> PCS -> proof -> PCS -> proof -> PCS -> device profile -> device link
where PCS is the profile connection space.  The CMM obviously would
normally mash these together in a single transform.  This is what
occurs in lcms.

This implies that the output color space for the device link
profile must match the color model for the real target device and
that the input color space for the device link profile must match
the color space specified for the device profile.
Still need to do some additional testing and update documentation.