ps2write - fix uncoloured Pattern spaces with a complex colour space
Certain kinds of colour space, eg /Separation, can have a function which
converts to the alternate space. If one of these was used as the colour
space for an uncoloured pattern then the function was not properly loaded
from the object defuinition, leading to an invalid colour space declaration.

This tests the initial space to see if it is /Pattern and if it is, checks
to see if the pattern is uncoloured (the array contains more than /Pattern).
If so we test the colour space and, if it is an array, process the array as
a colour space, which properly dereferences all the elements.


Expected Differences:
09-47a.ps, 09-47b.ps and 18-02b.ps should now work correctly.