pdfwrite/ps2write - Properly validate BitsPerComponent for images
Suprisingly (to me) the BitsPerComponent value for images is not validated
at the interpreter (language) level, but is handled in the graphics library.

pdfwrite and ps2write were not validating the BitsPerComponent, other than
to check it was 8 or less, and so could create invalid images in the PDF or
PostScritp output.

This commit adds the missing checks.

Expected Differences
12-07a.ps - ps2write no longer produces an invalid PostScript file
12-07a.ps - pdfwrite no longer produces an invalid PDF file
12-02.ps - ps2write no longer produces an invalid PostScript file
12-02.ps - pdfwrite no longer produces an invalid PDF file