Bug 692764: check Freetype is new enough
We must use Freetype v 2.4.0 at the earliest, so configure will now check that
if being asked to use a shared Freetype lib.

Add a configure check for Freetype shared lib that doesn't depend on pkg-config.

Move the warning about falling back to the deprecated AFS code to the end of
the configure output.

No cluster differences.