ps2write - Properly handle masked images with interpolation
ps2write converts type 3 and 4 (masked) images into an image and clip
combination. The clip is created from the mask, and the image is rendered
to a memory device. Note that the memory device canvas is just large enough to
contain the image.

The image is drawn as a series of rectangular fills, and our 'local conveter'
device shifts these from the original page location to the correct (relocated to 0,0)
position in the memory device.

However, if interpolation is true for the image, then we don't get a series
of rectangular fills, we get a 'copy_color' instead which the converter
device didn't handle. Adding a copy_color method which properly translates the
image position solves the problem.

Expected Differences
Progressions in Bug691210.pdf, 12-07B.ps and 12-07C.ps