ps2write - When converting shaded fills to image+clip, preserve insidenes rule
ps2write cannot preserve a fill with a Shading pattern, so it converts these
to an image and a clip (the path to fill is converted to a clip path).

However it did not preserve the 'insideness' winding rule which meant that
both eofill and fill were converted to 'clip' when eofill should be converted
to eoclip.

Expected Differences
fts_31_3110.pdf
fts_31_3115.pdf
fts_31_3118.pdf
09-47A.ps
09-47B.ps
09-47C.ps
09-47D.ps
09-47E.ps
09-47G.ps
09-47H.ps
09-47I.ps
09-47J.ps
09-47K.ps
09-47L.ps
09-47M.ps
18-02A.ps
18-02B.ps
18-02F.ps
23-12W.ps

All show progressions with ps2write