Shell script to convert HPGL/2-RTL to PDF.
Aside from the obvious purpose of converting hpgl/2 files to pdf, the
script demonstrates how to use the bounding box device to find the
extant of the graphics in the gl/2 file and then use those bounding
box coordinates to defiine a plot size for the file so there is no
white space margin.  This is commonly needed as the plot size command
in GL/2 files specifies a plot much larger than needed.