PJL support for the resolution variable.
The PJL resolution variable is now properly supported, for example "PJL
SET RESOLUTION = 300" will see the resolution for the next job to 300
but will be overridden if -r is given directly on the command line.
Note the PJL resolution is only one value, there is no way to specify
asymmetric resolution as there is with the -r option.