Fix pdfwrite and other vector devices for changing filename and separate pages
Previously vector devices did not allow changing OutputFile and (most) did not support the %d
OutputFile to generate single page files (pswrite was the exception). The function used by
pswrite was generalized and renamed gx_outputfile_is_separate_pages for general use. If this
returns true vector devices will close and reopen after each page. The PageCount will then be
used in the formation of the filename for the next page.

Since vector devices open an write the OutputFile when the device is opened the file will be
written even if there is no input, and if a format specifier is used in the OutputFile, then
an extra blank page will be written (maybe this will be fixed later).

For example, now pdfwrite can be used in server mode by changing the OutputFile device param
which will cause the current collection of pages to be written to the PDF file, and a new
PDF will be started. This is most useful from gsapi calls, but creating two PDF's from one
invocation of gs via the command line would be:
    gs -sDEVICE=pdfwrite -o tiger.pdf examples/tiger.eps \
     -c "<< /OutputFile (colorcir.pdf) >> setpagedevice" \
     -f examples/colorcir.ps