ps2write - alter /NumCopies handling.
Previously we mimicked /NumCopies and /#copies by using copypage. This is
unreliable because copypage differs on level 3 devices and won't work as
expected. Also it breaks DSC compliance and is probably undesirable with
other PostScript processors such as CUPS.

This commit alters the behaviour so that we use copypage only if the flag
DoNumCopies is true *and* we ProduceDSC is not true. This changes the
default to match pdfwrite and what I think is most logical while preserving
the ability to use copypage in environments where its use can be supported.

Expected Differences
Any files using /NumCopies or /#copies will behave differently, producing
fewer (blank!) pages than before. In our test suite I think this only
affects the Quality Logic files. In particular 268-03.ps and 29-07b.ps