ps2write/pdfwrite - Don't arbitrarily change VVECTOR0 flag in type 1 info
Way back in 2003 code was added 'in lieu of proper CDevProc processing' to
alter the behaviour of pdfwrite. In particular a flag was set which always
causes pdfwrite to handle width chnages.

However, we do now handle CDevProc correctly, and the arbitrary assignment
of this flag causes serious problems for pdfwrite & ps2write in the
(admittedly rare and unreasonable) case where the FontMatrix of an already
used font is modified.

Expected Differences:
14-12.ps
16-05.ps

Both now work better with ps2write and pdfwrite, though not 100% correct