Fix for bug 692733.  Pattern clist improperly cleared.
The pattern code was performing an initial clear of the pattern.  This is OK
when the pattern is not a clist since the mask was not drawn in when this
fill occurs (seee gx_erase_colored_pattern).  Unfortunately this rect fill
was getting into the pattern clist which is should not.