Fix Bug 692343 - ignore invalid padding length bytes.
AES streams contain padding at the end to bring them up to 16
byte chunks. The last byte of the plaintext is supposed to be
a value between 1 and 16 telling us how many bytes to ignore
from the last one? (Why 16? 15 would make more sense).

If the padding byte was out of range we would previously have
thrown an error. Here we change to just warn and set the
padding length to 0. This means we'll err on the side of making
the stream too long, which is better than the alternative.