ps2write - do not allow Widths aray to alter /.notdef width.
The way the Widths array works, if the Encoding has a /.notdef entry then
the width of the /.notdef is altered to that width.

If we have an earlier re-assignment, this breaks the assigned width.

The file catx1490.pdf relies on a ./notdef to draw a 'box' (crazy but
that's what it does). If we redefine the width tehn it is incorrect.

Expected differences
Catx1490.pdf
Bug691221.pdf
Bug689757.pdf
THe latter two are minor differences, but slight progressions