ps2write - correctly convert ICCBased *image* colour spaces
Previously we handled ICCbased spaces by pretending they were device spaces
with the same number of components as the ICCBased space. This works
suprisingly well, but fails utterly for Lab spaces where images come out
completely incorrect.

We now convert the image data to the current device space. We do not yet handle
vector colour spaces, this will be handled by a future rewrite of the colour
handling to permit much better control over colour space handling.

This change causes quite a number o differences in the regression suite.
These fall into one of 4 causes:

1) Colour space conversion; previously ICCBases spaces with 1 or 4 components
would be handled as grey or CMYK respectively. They are now converted to the
device space (RGB by default). This causes some colour changes, especially
in halftoned output with a 1 component space. These are not incorrect as can
be seen from the contone output

2) There are some minor shifts in colour caused by correctly handling the
ICCBased space, in general these are progressions (especially the Altona
suite).

3) Handling of Lab colour spaces. These now work correctly where previously
they did not. Especially catx4929.pdf and IA3Z3476.pdf

4) Honouring rendering intents. Files hwich used a rendering intent now are
correctly converted. cf Bug #691926, comment 3.

5) There seems to be a possible problem with converting some CalRGB spaces,
the resulting output seems 'blue' with respect to the Acrobat otuput. In the
case of 1 file which contains scanned images of books, it is clear from the
original book covers that the output is incorrect. Michael suggest this may be
a White POitn problem. A separate bug report will be filed for this. Ghostscript
rendering of the original matches the ps2write output.