Bug 690473: scale substitute fonts to fit in Widths.
Compare the width of the glyphs in a substitute font with those in the PDF
Widths array, and scale the font down to fit better, if necessary. This
does not add scaling up the font.

This changes a number of cluster files - most are progressions, some are just
different (no worse, no better), none that I consider regressions.