Improved page size matching in the CUPS Raster output device
The tolerances when comparing the page sizes in the document with the
sizes in the PPD file are changed to 1% for the short dimension of the
page and 5% for the long dimension. The 1% for the short dimension is
more or less the same as the 5.0pt used before (on letter or A4
paper). The 5% can cover the half inch deviation which HP puts into
the sheet length for their duplex paper sizes on ink jets.

If there is no matching paper size in the PPD and therefore the paper
size is considered custom, the page will be rotated in the case that
it does not fit into the maximum custom paper size (= printer
dimensions) but fits rotated.

With these measures, pages with slightly incorrectly defined size or
slightly incorrectly defined sizes in the PPD file (like the different
entries for one and the same paper size in the PPDs of HPLIP) will
find their size in the PPD and so rotated if needed.

This avoids cut-off print outs or even crashes of the printer
(https://bugs.launchpad.net/ubuntu/+source/linux/+bug/917148) due to
landscape pages not getting rotated.