Remove last vestiges of halftone cache.
Back in 2002 code was committed (as part of the DeviceN work) that
removed the halftone cache from the image state. The last vestiges
of this were functions:

 gx_check_tile_cache_current
 gx_check_tile_cache
 gx_check_tile_size

All these were hardwired to return a constant value, but the calls
to them were never removed from the code.

Here, we remove those calls. No differences should be given.