pdfwrite - fallback to 'LeaveColorunchanged' if we cannot deal with color conversion
Some of the PDFSETTINGS collection of settings set the ColorConversionStrategy
to spaces which may not be possible (in the current architecture) to convert
every colour space into. For example ICCBased spaces. This causes a fairly
incomprehensible error to be thrown.

In future we intend to do a better job with colour taking advantage of the
built-in colour management, but for now we simply choose to change the
ColorConversionStrategy to 'LeaveColorUnchanged' and emit a warning.

Bug #692030 and Bug #692108

No Expected Differences as this is not tested by the cluster