Bug #688227 Alter the PostScript 'show' operators so that the show enumerator is in local VM
As described in the bug, it seems to be technically possible for a local font to be
referenced from a show enumerator. If the enumerator is in global VM this
causes a garbage collector crash.

This code modifies the PostScript 'show' family of operators so that the
enumerator is always allocated from local VM, avoiding the problem.

There is no specimen or method of reproduction supplied in the bug report so
this is rather speculative, however it runs a cluster test without problem.
If it does cause any problems we should simply revert this change.

As part of this change it was noticed that there are a number of procedures
defined in gschar.c which are apparently dead code and these have been removed.

No expected differences.