C portion for making use of output intent ICC profile.
This commit puts into place the c-code for making use of an
ICC profile stream from the PDF interpreter for setting of the
output intent.   This is done through the use of zset_outputintent.
This function will make the various settings of the icc profiles
in the device and in the icc manager based upon the process model
of the target device and the icc profile of the output intent.  It
will be up to the interpreter to make the call to  .set_outputintent
with the proper output intent ICCBased dictionary on the top of the
stack.  This should be done in the interpreter after the device has
been set up and all the default ICC profiles are in place.  The
interpreter should make the call if we have specified -dUsePDFX3Profile.
The first output intent that is present in the file should be used,
unless someone specifies a different intent through the use of
–dOutputIntent=x   .  In this case, the x+1 output intent (if it
exists) will be used.  Note that if the particular output intent does
not include an ICC profile entry (DestOutputProfile), then the output
intent will not be used.  In this case, we will eventually want to
specify what the OutputConditionIdentifier was for that particular
intent if it was missing the DestOutputProfile.