Fix interpolation of indexed colorspaces.
WMB1102317A01.pdf (supplied by a customer) shows a problem with
Interpolation of Index colorspaces.

Specifically, we have an Indexed DeviceGrey space (which is, in
the new world order, an Indexed ICC space, with 1 component).

At line 678 of gxiscale.c we correctly detect that pcs->type->index
is Indexed, and therefore use the base space as the 'actual' space.

At line 683 we find the concrete space (which in this case is
exactly the same space). We then decide whether we are working in
an icc world, or a device_color world by checking for the presence
of icc data. The old code used to (incorrectly) check pcs (the
indexed space), when it should really have checked the concrete
space.

The remaining changes in the code are purely tidying:

 * avoid calling concrete_space a second time (when pconcs already
   holds the answer)
 * pass pconcs as an arg when calling pconcs->type->remap_concrete_color
   rather than pactual_cs. At that point pactual_cs and pconcs must
   be identical, but it's nicer to be consistent.

This produces 3 differences in the cluster tests, all in the cups
tests. These can be reproduced locally using tiffsep.

 comparefiles/148-11.ps.cups.300.1
    Definite progression

 PDFIA1.7_SUBSET/CATX1028.pdf.cups.300.1
    Looks like progression to me

 ps/ps3cet/12-07B.PS.cups.300.1
    Definite progressions on page 12-15. A problem still exists on
    page 12 though, but that's unaffected by this change.