Fix for bug 692865.  Make XPS interpreter use ICC color spaces for default
The XPS interpreter was still using Device color spaces for its default
color spaces.  This fix has it use ICC color spaces by default which
is required by the spec.   This revealed an issue with the color space
handling of the opacity masks.  These were getting drawn in sRGB color
space but the transparency group is linear gray.  Since we are drawing
the opacity mask as a luminosity mask that is gray we ended up
mapping through a nonlinearity.