Movement of code for setting ICC system/user params into graphics lib.
The code that set the system/user params related to ICC profiles was
contained in the zusparam.c file.  This file is not used with the
other languages (e.g. xps and pcl).  The functions that do the actual
work have been moved to gsicc_manage reducing any code duplication for
the other languages.