Support HPGL style path handling.
HPGL handles paths slightly differently to the postscript imaging model.

When a path is filled, successive moveto's are treated as linetos.
(i.e. the 'place we close the path to' is left at the first moveto,
and the area remains fillable). Stroking is unaffected however.

To model this in Ghostscript we add a new path segment type 's_gap'.
The filling code treats this as a lineto. The stroking code is updated
to not stroke such edges (and not to break the subpath at this point).

We add a new parameter to the imager state (hpgl_path_mode), new
accessor functions (gs_sethpglpathmode, gs_currenthpglpathmode),
and new postscript operators (.sethpglpathmode and .currenthpglpathmode).

If hpgl path mode is set to a non-zero value, then path construction
treats movetos in an open subpath as gaptos.

Still to do:
  * Update PDF write to spot such paths and to convert them as
    appropriate when writing out.