User params reset issue of icc profile strings
If the user params end up getting set during a vmreclaim
while we are in a swapped icc profile situation when
processing a softmask we were ending up freeing profiles
that should not have been freed.   This issue was masked
by a check that was disallowing profiling settings in
the manager once they were already set.  Also discovered
issues with a corner case where we have a softmask that
includes a pattern with a softmask.