Fixes bug 692894, orientation of pcl pages incorrect.
This is a case of the current code emulating undocumented behavior
incorrectly.  It appears the HP printers resets the print direction
upon receiving a new page size command. Previously we tried to emulate
this by not resetting the logical page orientation which is incorrect
but happened to print the tests we had correctly.