Change compression of the tiffsep device composite output to match the separations.
Previous to this commit the tiffsep device would always write out
an uncompressed composite file; the separation files were lzw
compressed by default and this could be changed via the -sCompression=
option.  Now the compression of the composite file is the same as
that of the separation files.

Fixes Bug 692907.