pdfwrite - first pass at PDF/A-2 output
This is the first part of making it possible to produce PDF/A-2b output from
pdfwrite.

The PDFA switch has changed from a boolean to an integer, where the value gives
the level of PDF/A compatibility. This has knock-on effects throughout the
C and PostScript code which has been revised to expect an integer instead
of boolean value.

When PDFA has the value 2 we no longer flatten transparency, and we write
'2' in the pdfaid field in the XMP metadata.

PDF/A-1b output still seems to work correctly, but it is unlikely that the
work so far is sufficient for correct PDF/A-2 output.

No differences expected as the cluster does not test PDF/A output.