pdfwrite - correction to PDF/A processing
The previous commit for PDF/A prcoessing did not properly check the PDF/A
level in the PostScript handlers, and treated any setting as being PDF/A.
In fact only settings greater than 0 shuold be checked, 0 means 'not PDF/A'

Add code in gdev_pdf_put_params to ensure that PDFA has a permitted value
(0->2) and signal rangecheck if not.

No differences expected.