Fix for memory issue in icc user params
The icc user parameters where improperly being set as being persistent
for the user params list when they were not.  In the case when we are
using the output intent profile, one of the default profiles is freed
and replaced by the output intent profile.  In this case, the default
profile contained the string that was used for the user params.  During
a VMreclaim, we would then go and restore the user params and
unfortunately this default one had been freed.