pdfwrite - cure memory leaks caused by 'aside' streams
When writing an object it can happen that pdfwrite needs to write another
object (Eg a ToUnicode CMap in a font), it does this with an 'aside' which
opens a Ghostscritp stream to contain the data.

The code to do this uses low level access to the stream code, but it did
not previously correctly free the memory used by the stream and its associated
buffer leading to memory leaks.

By closing the entire stream when we close the filters we get the stream and
all its memory freed, so we do that here to cure the memory leak.

Inspecting the four other cases in the pdfwrite code they appear to be OK.

No differences expected.