ps2write - work-around bugs in Brother printer CCITTFaxDecode filter
Raised on irc by Till Kamppeter, see Ubuntu bug :

https://bugs.launchpad.net/ubuntu/+source/cups/+bug/955553

After much work by Chris Liddell and Bruce Stough it transpires that at
least some Brother printers have a bug in their CCITTFaxDecode filter,
especially with small amounts of data.

Since the ps2write otuput for bitmapped glyhs (which is required when the
input is a CIDFont) always uses the CCITTFax filter, this led to corrupted
output from the Brother (Note this is a bug in the *printer* not ps2write)

This patch adds a new command line parameter 'NoT3CCITT' which disables
compression of bitmapped glyphs. It should only be used with printers which
have a problem with CCITTFax data, and in that case should also be used with
the EncodeMonoImages switch to disable compression of monochrome images. Eg:

-dNoT3CCITT -dEncodeMonoImages=false

No differences expected as these are not tested by our regression tests