Close the device before exiting PCL, XPS and friends.
The languages deallocated the memory associated with a device but
never did the associated gs_closedevice().  This might have some
effect on 692816 where indeterminate output was observed on stdout.
Without the close device call the "OutputFile" was never actually
closed (fclose), and so there might have been a missing flush.  All of
that is speculation, I wasn't able to reproduce the indeterminism -
but we want the closedevice anyway so here it is.