Bug 692850: FAPI: clamp fixed-point overflows instead of error.
This job ends up with an insanely big scale factor when trying rendering glyphs
from a font. Obviously that ends up too big to render to a bitmap in the
font renderer, so we build a path instead.

Previously, during FAPI path extraction, if the coordinates overflowed what we
can represent in our fixed point representation, give a rangecheck error. But
the AFS code seems to clamp the coordinate to something we can represent, and
carry on.

Do the same in FAPI.

No cluster differences.