pdfwrite - update CIDSet emission for PDF/A-2 compatibility
This completes the work for creation of PDF/A-2 files. Note that Acrobat X
preflight complains about the CIDSet we include in PDF/A-2 files. However I
believe that the current code is correct, and that the Acrobat preflight is
incorrect. The PDF-Tools validator is completely happy with our CIDSet.

I'm unable to fathom what it is the Acrobat preflight tool wants. The old
PDF/A-1 CIDSet is flagged as invalid, as is every premutation of values I
can think of.

In the absence of any example file to look at, or a definitive resource on
the subject, I'm comitting this and concluding the development. I will of
course be happy to work on the CIDSet emission if someone can bring more
information.

No differences expected PDF/A is not tested by the cluster.