Ensure FAPI uses appropriate glyph name for metrics.
Found investigating a customer 532 report: Ghostscript "augments" glyph names
under certain circumstances, and depending on the font type, FAPI may have to
remove that augmentation (mainly when disc based fonts are being used).

FAPI was then using the un-augmented glyph name to read the glyph metrics,
which caused the wrong metrics to be used because the metrics(s) dictionary was
created with the augmented glyph name in place.

Causes a difference in the cluster file Bug691031.pdf - this is
progression.