Bug 692950: further revise falling back to unhinted rendering.
In the FAPI/FT interface, we caught some errors from Freetype to do with the
byte code hinting in TrueType fonts - when they occur we retry the glyph
rendering with hinting disabled.

This commit expands that to *any* error from the byte code interpreter will
now cause us to retry unhinted.

(the core problem in this case seems to be scale depending operations in the
font program, which is supposed not to be permitted).

No cluster differences.