pdfwrite - when a Charstring is not found for a glyph use the /.notdef width instead of 0
Bug #692944 the file uses an embedded font which does not include a /space
glyph, but proceeds to use the space anyway. We were using a widht of 0 in
this case but the file programmatically spaces glyphs to a predefined size.
This was causing the PDF to be incorrectly spaced.

We now use the width of the /.notdef glyph instead for sizing calculations,
since this si the glyph which will eventually be used.

This exhibits a small difference in 30-06.ps but since this is testing
missing glyphs in a QL test file, I'm going to accept it as a difference in
order to improve the result with files of this type.