Add psdcmyk support to bmpcmp.
bmpcmp now reads psdcmyk files. It will read all n planes (cmyk +
spots) and search in them for diffs. When it comes to output
the differences are highlighted correctly in the diff image,
but only the cmyk planes are used to create the output images
(i.e. spots are not mapped down to cmyk currently).