Delay of obtaining handles from CMM for default profiles
Obtaining the profiles handles for all the default profiles in
the ICC manager eats up a tiny amount of time on startup.  For
some customers that start and stop gs over and over instead of
running in server mode, this startup time is an issue.  This
fix performs the initialization in a lazy manner, obtaining the
profile handle only when finally needed.

This commit also fixed some issues with images that are in the
LAB color space.