Update bmpcmp to (at least try to) honour the color mapping for psd spots.
We read the color info for the spots, and we try to apply it when mapping
down to cmyk. It's entirely possible that the color information needs
to be inverted or reordered or something, but I don't have anything
that can display psd files to check.

Talk to Michael about this when he returns. This is better than nothing.