Fix ref counting of CIEBased 'params' during alloc_restore_all with 11-19.PS CET.
The 'const' had to be broken to allow the pointer to be set to NULL when the object is freed
due to reference count going to zero. Debug build showed an error message due to ref_count
going below zero, but since this is referencing freed memory, can cause a segfault.